requirements.txt
aiosmtplib==2.0.1
alembic==1.10.2
anyio==3.6.2
bcrypt==4.0.1
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==39.0.2
distlib==0.3.4
dnspython==2.3.0
docopt==0.6.2
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.93.0
fastapi-mail==1.2.6
filelock==3.5.1
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
motor==3.1.1
passlib==1.7.4
pipreqs==0.4.11
platformdirs==2.5.0
psycopg-binary==3.1.8
psycopg-pool==3.1.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.27.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.6
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.9
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.13.1
watchfiles==0.18.1
websockets==10.4
yarg==0.1.9
